add go.mod file and corresponding dummy imports
This configuration is sufficient for 'go list k8s.io/...' to succeed.

It was produced by iterating on the following process:
• go mod tidy && go list k8s.io/...
• go get -m <missing dependency>

To the extent possible, all k8s.io versions are selected using v1.13.3
or kubernetes-1.13.3 tags

k8s.io/heapster requires special treatment: it is a dependency of
packages in the k8s.io/kubernetes module, but contains packages that
do not build using any recent Kubernetes release. We use a local
replacement to prune it down to only the package needed by
k8s.io/kubernetes.